Avoid spamming monitor logs when reconnecting to watchman
Summary:
I noticed that we spam the logs when we lose a connection to Watchman. [There's a backoff time of 4s](https://github.com/facebook/flow/blob/1f0fe32d8817be0796b7acb7e96a53bf490ac379/hack/watchman/watchman.ml#L544) for the first attempt, 8s for the second, 16s for the 3rd, and 32s for the rest. During the backoff, the `Watchman_lwt.get_changes` call is not blocking. This meant we basically had a busy wait loop spamming the logs and spinning the CPU.

Instead, let's yield for a second each loop.

Reviewed By: jbrown215

Differential Revision: D13107830

fbshipit-source-id: b2296a9ff6324a49b29bf7ec698f075edd6dd1b7