[FIX] Use a single constraint for all local exports of DeclareModule
Summary:
Before, we would create a constraint
```
ModuleT ~> ExportNamedT (..., SMap.singleton x t, ...)
```
for **each** exported local declaration `x` of a "declare module". This created big chains
of constraints for libdefs, leading to deep call stacks in flow_js and eventually
an exception when running flow.js on node/chrome.

This changes the above behavior to create a single constraint of the form
```
ModuleT ~> ExportNamedT (..., local_exports, ...)
```
where `local_exports` is the mapping from each *local* exported binding of a
DeclareModule to its location and type.

Using a map here is safe because the type recorded already accounts for multiple
declarations of the same name.

Reviewed By: samwgoldman

Differential Revision: D13212805

fbshipit-source-id: f5b216894c6222b077099d5e8ba191ea9236c9ca