[react] Move get_config to toplevel
Summary: This is an unfortunate diff. A lot of the functionality hidden in `run` will be useful for `AbstractComponent`. However, it relies on functions in `flow_js.ml`. This file can't depend directly on flow_js.ml because it will create a dependency cycle. That means that any functions that get moved to the toplevel in this file will need the flow_js functions passed in explicitly. This moves all of the functions necessary to get `get_config` into the toplevel. `get_config` will be used for typechecking `AbstractComponent`

Reviewed By: samwgoldman

Differential Revision: D13125070

fbshipit-source-id: c016979d1931cc02fb14316f256a041e0ce3fa60