[react] AbstractComponent ~> TestPropT
Summary:
To handle TestPropT and friends (Get/SetPropT Get/SetElemT), we propagate a (SFC | React$Component) ~> TestPropT constraint. This means that we handle TestPropT and friends as we handle React$ComponentType as it is defined today.

The reason we do this at all is because people seem to want to be able to access displayName or name on React components. A common pattern is `const x = Component.displayName || Component.name`

Reviewed By: samwgoldman

Differential Revision: D13003366

fbshipit-source-id: 2e53438b7a8015ea45d43078a6b74d82348ed141