[react] Use config instead of props in ClassT ~> AbstractComponentT, make checks contravariant for configs
Summary:
Now that the first argument to AbstractComponentT is config instead of props, we should correct the already landed that assumes it would be props.

In order to check the configs contravariantly, I added a polarity argument to ReactConfigMerge which lets flow know how to flow the result of the ObjKitT.

I also make all of the config checks contravariant here, including the checks in AbstractComponent ~> AbstractComponent.

The updated test case will look funky. Remember that the first argument is config, so any defaultProps are optional.

Reviewed By: samwgoldman

Differential Revision: D13125135

fbshipit-source-id: 8f9431a2176a988b3b7cd6e59a7ac75267a59994