Always color coverage command output
Summary:
Nobody passes `--color` so they can end up with the contents of their file reprinted as-is. That's what currently happens if the output isn't a tty.

With this change, we always color the output if `--color` is passed. This makes the `coverage` test output a bit harder to read but it shows up right if you `cat` it in a terminal which supports coverage.

This makes it so that we can actually test the `--color` output.

Reviewed By: gabelevi

Differential Revision: D13221253

fbshipit-source-id: 31dfebe416afa27ee1aa4a705685728fd0e3fecc