[react] createElement for AbstractComponent
Summary:
This implements createElement for AbstractComponent. The implementation here is a bit hacky, since we already have access to the config.
I explain this in a comment inline.

The reason I didn't just rewrite this entire function to use get_config and then compare the incoming config object with the config type is
because there is some special handling we have for children, keys, and ref that is handled in `object_kit`.

That said, this is something that I want to refactor in the future.

Reviewed By: samwgoldman

Differential Revision: D13221333

fbshipit-source-id: a41e86a4769abf21006d352909cbab4dddb05fa8