[hack] Convert `hh_client` to Lwt
Summary:
I did this to check to see if it was possible to convert the entire codebase to Lwt. It wasn't, but we might as well check this in.

If we converted `hh_server` to Lwt, we could remove the `Process`, `Daemon`, and `Future` modules outright, and significantly simplify the implementation of the worker system. However, it's difficult to do incrementally, and probably doesn't add a lot of value to do at this point.

Reviewed By: ljw1004

Differential Revision: D13197032

fbshipit-source-id: 62fbce0f5a10cd9937baea912c3695662d32a83a