architecture option
Summary:
Later in the stack, a new "types-first" architecture will co-exist with the current "classic"
architecture.

We expect the "types-first" architecture to be in development for some time, during which it may not
be externally testable. Thus, a hard-coded setting of "classic" suffices for now. Before long we
will add an external flag to control this setting.

Reviewed By: panagosg7

Differential Revision: D13254190

fbshipit-source-id: adb0402992bdd72efb026588c72a19dbdfa88016