[react] Add AbstractComponent to ComponentType and ElementType
Summary:
Adds AbstractComponent to the ComponentType and ElementType unions.

We can completely replace the ComponentType union with `AbstractComponent<Config, any, any>`, since ComponentType was already conflating config and props.

Re-recorded tests have the same error messages, but point to different libdefs. the only case this isn't true is the tests in tests/react, where the messages changed slightly, but the same errors are found

Fixes: https://github.com/facebook/flow/issues/7220
Fixes: https://github.com/facebook/flow/issues/6901

Reviewed By: samwgoldman

Differential Revision: D13005204

fbshipit-source-id: f573a98cbdd0d72ec57a801625a44ca1b49b886d