[react] React$Config
Summary:
Adds a type-level function React$Config<Props, DefaultProps> which calculates the config type given the props and defaultProps.
In a stacked diff I will add this to lib/react.js and re-record tests.

We use the EvalT infrastructure here, as is done with the rest of the React destructors. Added tests include the proposed "best practices" way to write HOCs.

Reviewed By: samwgoldman

Differential Revision: D13263769

fbshipit-source-id: b785d616d1ec902213990bf45e740333b8192db2