[hack] Fix random id bug when it contains slashes
Summary:
Hack uses random ids in many places. Some of them are used for filenames, but turn out to have `/`es in them which causes random occasional failures.

Gets rid of random ids with `/`

Impact on Hack is that the random ids have reduced entropy. One way we could address this is to add 2 characters that are valid in filenames.

Reviewed By: ljw1004

Differential Revision: D13309389

fbshipit-source-id: 9c01c7cdaeeae5f83e938b47422046076521726c