When --from is omitted use parent process' cmdline
Summary: Lucian gave me a tip that Tatiana had built a nifty little library which can read some info from `/proc/<PID>/stat` and `/proc/<PID>/cmdline`. This makes it easy to grab the parent process' cmdline when on a system with `/proc` available. Hopefully this will improve the logging for commands which omit the `--from` flag!

Reviewed By: ljw1004

Differential Revision: D13284665

fbshipit-source-id: 634a7e60ec51f918c8ab828498c8095dc21fbebd