Remove unnecessary copying from the copy_of_context call
Summary:
The places that call this function only care that context optimization /
clear_intermediates does not affect the copy. Those modifications only affect
the mutable fields of the sig context, not the underlying mutable things.

Before this change we were copying the entire tvar graph, for example, which
was unnecessary. We also copied the type tables, but those are part of the
local context and not the sig context, so neither context opt nor
clear_intermediates even touch it.

Reviewed By: pakoito

Differential Revision: D13217949

fbshipit-source-id: 08ad55c086ac1f15ad87b8facd434989cde056cf