[ast] kill "assignment patterns"
Summary:
"assignment patterns", like `pattern = expression`, aren't a real thing in the spec.

we use them to represent defaults for array and object destructuring, and function params. but since the `Pattern.t` variant contains `Pattern.Assignment`, you can construct bogus ASTs like `let (foo = bar) = baz`, which don't actually parse. further, we have a bunch of places that have to `failwith` or otherwise error when they see these patterns in invalid places.

so, I added `default` fields to `Pattern.Array.Element`, `Pattern.Object.Property` and `Function.Param`. this is nice because it (a) covers the valid uses of `Pattern.Assignment` and (b) forces everyone to handle defaults, whereas before you had to remember that a pattern might be a `Pattern.Assignment` which is actually a default. Destructuring is significantly more complicated because of `Pattern.Assignment`. this might help.

Reviewed By: pakoito

Differential Revision: D10853973

fbshipit-source-id: eec9be42da088121e814221e4d020e98655da7bd