[hack] Don't send file edits in parallel
Summary:
Sending file edits in parallel is not allowed; all RPC calls must be serialized.

This commit adds a check to `ClientLsp.rpc` to make sure that you're calling it in a serialized fashion. (There is an `Lwt_mutex` library, but I don't use it because it's probably a mistake for the user to have tried to parallelize these RPC calls so long as the server doesn't support parallel RPC calls, so let's fail fast in that case.)

Calling `rpc` in a non-serial fashion may cause essentially undefined behavior.

Reviewed By: ljw1004

Differential Revision: D13355252

fbshipit-source-id: 2987c61b8798c7c6f7b1f8d5c575df0216dd6ae0