[normalizer] Use flag to opt for union/intersection optimizations
Summary:
Earlier the normalizer would optimized unions/intersections as they were created. This
could be expensive in some cases redundant, since different services may have different
needs in that respect:
- coverage does not need types that are deeply optimized since the types themselves are
not exposed
- type-at-pos, suggest, dump-types, etc, OTOH do.

This change:
1. makes this an optional flag passed to the normalizer, and
2. does the optimization pass once on the final result.

Since we do not optimize types when computing coverage, we modify the process that
determines. A type `T` is not covered iff one of the following holds:
* `T` is `any`
* `T` is `empty`
* `T` is `t1 | ... | t2` and *all* of `ti` are uncovered
* `T` is `t1 & ... & t2` and *any* of `ti` are uncovered

Reviewed By: samwgoldman

Differential Revision: D13326834

fbshipit-source-id: 19c4b820635e6f1adca90375634d2eb76f7b4d39