[parser] skip invalid object destructuring props on error
Summary:
methods, getters and setters are not allowed in object patterns, and we error in those cases. but we tried to recover by creating an "expression pattern" containing a function, except that patterns can't contain functions either and our AST is too loose to prevent it.

this diff just skips these properties in the resulting, error-tainted, AST.

Reviewed By: gabelevi

Differential Revision: D13220206

fbshipit-source-id: 5791cded52af4d338c8c5ab6763a80197c7b6f5a