Clean up handling of unchanged focused files in recheck
Summary:
Long ago, recheck was only called on files which the file watcher reported as changed. Those were simpler times.

Then lazy mode came along and recheck could also force files to be focused.

Then file hashes came along and we could notice that a file was unchanged on disk. However, if we were trying to mark an unchanged file as focused (and it wasn't focused before), we would pretend like that file was changed and freshly parsed. This was kind of hacky, but worked.

However, I'm adding more complexity to recheck and want to clean up this hack. Rather than pretend that unchanged focused files are changed, I'm just going to keep track of them separately.

Pros:

* It's less hacky
* We don't need to update the dependency graph for these unchanged files
* It makes it easier for me to also ask recheck to force files to be marked as dependents or dependencies (currently you can only force files to be focused).

Reviewed By: avikchaudhuri

Differential Revision: D12847068

fbshipit-source-id: 5d37a28c8bac63dd71e519c973143e5b15662189