Recheck with a CheckedSet instead of a set of files to focus
Summary:
Currently recheck can only force certain files to be focused. I need to make it so it can force files to be dependents or dependencies too. So instead of taking a set of files to force focused, I'm changing it to take a CheckedSet of files to force.

This diff updates the implementation of `recheck` to take a `CheckedSet`. There is, therefore, new logic to handle if we're forcing files to be dependents or dependencies

Reviewed By: avikchaudhuri

Differential Revision: D12870144

fbshipit-source-id: 4bb2ad4c55f8cd2eca1ddc34ee7efd1d50a1e508