Optimize Dep_service.dependent_files no-op
Summary:
`Dep_service.dependent_files` calculates the direct and
transitive dependencies of some root files. It is also passed a set of
candidates which may be dependents. The calculation can be
O(candidates), and for very large repos this can take multiple seconds.
So in case we have no root files we can short-circuit that behavior.

Reviewed By: avikchaudhuri

Differential Revision: D12987427

fbshipit-source-id: 4d44a8314b437356b946f0400dded44e4596836e