[React] Add use-types to model react functional components instead of using any
Summary:
Previously in order to model props flowing to and from React functional components we were creating an `any`-valued function in order to ensure that the functional component had the proper form as well as correctly-typed props. This would have unnecessarily propagated `any` in many places around our React model and tainted a whole bunch of stuff unnecessarily. This should fix that for function/callable object components. The same work on class components will be in a followup diff.

This also has the upside of fixing some non-errors being reported as a result of function arity.

Reviewed By: jbrown215

Differential Revision: D13386555

fbshipit-source-id: 05043b81825b1c9631331f3761da1be719384550