nominal ids are locs
Summary:
Class ids being fresh ints causes correctness problems in types-first, and also lead to
complications in sig hash computation. Zooming out, they are needed for nominally distinguishing
classes. The location of a class name works fine for this purpose.

Not sure what the performance implications of this change will be. We'll find out.

Aside: Some other kinds of types, like polymorphic types, also reuse the nominal id machinery. I
believe this is a partial abuse of the name, but I'm not confident that the expected behavior
doesn't overlap with some aspects of nominal typing. I have left that machinery intact for
now; perhaps we can revisit by renaming the concept to something else and tightening up its
associated API.

Reviewed By: samwgoldman

Differential Revision: D13395461

fbshipit-source-id: fe6469e713955719081256357f3a0366cad8674c