Offset utils
Summary:
I would like to remove `offset` from `Loc.position`. An early experiment showed that this resulted in a large decrease in memory usage. However, some places still use offset, so we need a replacement.

This introduces a utility that builds a table of `O(n)` size, which takes `O(n)` space, where `n` is the size of the program text. Then, offset lookups into this table are `O(1)`.

Unlike the offsets in locations, the offsets calculated by this utility are the number of bytes to offset into the file, not the number of characters (this matters in the case of multi-byte characters). An offset for the number of characters is next to useless (at least without a string type that natively supports multi-byte characters) since you would still need to perform a linear scan in order to count multi-byte characters as one character and arrive at the intended character. I have already found one place (the `coverage` command) which was using offsets and incorrectly assuming that they represent the number of bytes to offset into a file.

Reviewed By: gabelevi

Differential Revision: D13246769

fbshipit-source-id: 81fb9969a9cdc281e31f7a442a55a1353e46c4a7