Remove `offset` from file sig tests
Summary:
As with other diffs in this stack, the goal is to remove `offset` altogether for the memory savings. This replaces uses of `offset` with `Offset_utils`.

We probably compute the offset tables more than we have to, but it's just a test so it's not performance critical, plus the examples are pretty small so it should be really fast to compute the tables.

Reviewed By: samwgoldman

Differential Revision: D13259591

fbshipit-source-id: 49ebcc775086078a3896427940186d8acf26c182