[Normalizer] Bypass TypeDestructorTriggerT when extracting RTypeAlias
Summary:
Departing from the general `TypeAppT` pattern, `TypeDestructorTriggerT` may imply a
type application itself without the presence of `TypeAppT` constructor. To avoid the pitfall
of missing an implied type argument we avoid inferring type aliases when the underlying
type is a `TypeDestructorTriggerT`.

The effect of this type ought to have been computed and produced the necessary lower-bounds.
This is why we already ignore this type in the main normalizer pattern match (this case returns
`Ty.Bot`). This diff extends this treatment to the early catch case of type-aliases.

Reviewed By: avikchaudhuri

Differential Revision: D13388387

fbshipit-source-id: c6a018b6051850cc33b786286428ba881bb07b27