improve hashing of use ops
Summary:
Constraint hashing for avoiding non-termination involves, in particular, hashing use ops.

However, the existing mechanism doesn't make a lot of sense to me: as we start with a toplevel
constraint and recurse deeper, we keep only the deepest use op in the hash, which means that we
might report multiple versions of the same error when we reach it via different paths from the same
toplevel constraint.

Instead, we can hash use ops to their toplevel use op. This not only has a more meaningful
semantics---errors are grouped by source-level operations---but might also help perf by cutting
redundant work.

Reviewed By: samwgoldman

Differential Revision: D13430430

fbshipit-source-id: a2d00a8abfa0189f78ac538d6ed22184a283956c