remove any generated when refining function from mixed
Summary:
Prior to this diff, in the following example `x` was refined to `Any`:

  function fun(x: mixed) {
    if (typeof x === "function") {
        (* do something *)
    }
  }

This is unsound, because while we know that `x` must be a function we don't know anything about its type beyond that. This modifies the refinement code so that `x` is refined to a function of type `empty => mixed`. This function can never be applied, which makes more sense and is more sound than allowing the function to take any arguments. However, properties like constructors and prototypes that are common to all functions can still be accessed.

Reviewed By: panagosg7

Differential Revision: D13123906

fbshipit-source-id: 7640e0d6df93c38e2e428de79add32ce95b47264