State readers
Summary:
In the future, we will run some commands in parallel with rechecking. The general idea is that the commands will read the oldified data.

This will be enforced using "State Readers". Init and recheck will be given a `Mutator_state_reader`. Commands will be given a `State_reader`. Code which is shared by both init/recheck and commands will be given an `Abstract_state_reader`.

Reading from shared memory with a `Mutator_state_reader` is basically like reading from shared memory today. You just read directly from the shared memory.

Reading from shared memory with a `State_reader` will generally check to see if we're in an init/recheck. If so, we look at whether oldified data is available, and if so we use that. Otherwise, we read directly from the shared memory.

Reviewed By: samwgoldman

Differential Revision: D13416796

fbshipit-source-id: 07791ad5d16ccdba76bbb93f6cddd27a8d648cb6