Separate worker "done" from worker "idle" metric
Summary:
Currently the "idle" metric measures two things: (1) time that the worker is
waiting on a job and (2) time that the worker is unused because there are no
jobs left, but some other workers are busy.

It is useful to separate these metrics because the strategies to optimize them
are different. For (1) we want to send jobs as quickly as possible. For (2) we
want to ensure that jobs are bucketed well so work is distributed evenly.

Reviewed By: gabelevi

Differential Revision: D13471705

fbshipit-source-id: ec287981b761ce43743ba1cc901fe38bc0a01e80