Fix Shape errors referring to wrong objects when properties are absent
Summary:
Previously the following code
```
class Empt {}
let _k : $Shape<Empt> = { x : 4 };
```
generated this error:
```
Cannot assign object literal to `_k` because property `x` is missing in object literal [1] but exists in `Empt` [2].

   test.js:81:25
   81| let _k : $Shape<Empt> = { x : 4 }; // flow error
                               ^^^^^^^^^ [1]
References:
   test.js:81:17
   81| let _k : $Shape<Empt> = { x : 4 }; // flow error
                       ^^^^ [2]
```

This is clearly wrong. `x` does in fact exist in the object literal, and not in `Empt`. This fixes the error reporting.

Reviewed By: panagosg7

Differential Revision: D13454042

fbshipit-source-id: 58078ce9a6a403e8ec00a842ddce068fe6ad1091