[hack] Server can send messages to monitor
Summary: The pipe between monitor and server already exists, but it's used only in one direction (incoming client FDs are sent through it from monitor to server). I'm going to use the other direction to send progress messages, so that monitor can know more details about the server than "it's alive / it exited with exit code X". This information can then be relayed to clients, even when the server itself is busy.

Reviewed By: ljw1004

Differential Revision: D13448065

fbshipit-source-id: d06f450fb8d28a711c0283111726dbec8ba734e4