[hack] Monitor can be queried for most recent server status
Summary:
One of the advantages of having the monitor, is that it's less busy than server (because it's simpler and written with this intention from the scratch, while server has a lot of "big blocking batch tool" legacy.).

We can use this responsiveness to offer insight into what server is doing while server is not able to tell it itself. This diff adds RPC command for that.

Reviewed By: ljw1004

Differential Revision: D13448063

fbshipit-source-id: 906b72e1010e35bf68e6d1b39299973e60648a4c