signature ast makes dependency graph
Summary:
Step (2) of types-first (see previous diffs) processes sig asts according to their file sigs, but
the parallel scheduling of this step still follows the dependency graph induced by asts and their
file sigs. In particular, large cycles in that dependency graph continue to cause poor parallelism
/ CPU utilization.

This diff computes the dependency graph induced by sig asts and their file sigs. This dependency
graph has fewer edges (the exported types of a module might depend only on a subset of its
imports); in particular, this means that there are smaller cycles in this graph. We schedule step
(2) using this dependency graph. This speeds it up nicely.

As with the results of parsing, we produce and consume the new dependency graph only in types-first.

Reviewed By: nmote

Differential Revision: D13254340

fbshipit-source-id: 82641a783f1b468a9e47cb1efe09c45ebc34cb9d