signature ast goes mainstream
Summary:
During parsing, a signature ast and its file sig are now recorded along with the ast and its file
sig. This information is exploited in a new architecture with better scalability characteristics.

In this "types-first" architecture:

(1) The signature ast and its file sig are also stored in new heaps.

(2) The current "merge" phase is repurposed to merge signature asts with their dependencies, as
recorded by their file sigs.

(3) A final "check" phase then checks asts against their dependencies, as recorded by their file
sigs.

Roughly, (2) and (3) together should do similar work as the current "merge" phase. However, only (2)
imposes the same restrictions on parallelism as the current "merge" phase (it must proceed in
dependency order), but in comparison it does far less work. On the other hand, (3) imposes no
restrictions on parallelism. Thus, we expect a significant perf win with this architecture.

Note that this diff still schedules the work in (2) according to the dependency graph induced by
programs; thus, parallelism is exactly like the current "merge" phase. This will be optimized later.

Reviewed By: nmote

Differential Revision: D13254261

fbshipit-source-id: 5d3b46c70c644fd2ef4e36ff545490d8881219b8