peak memory / time tradeoff
Summary:
Step (3) of types-first (see previous diffs) is embarassingly parallel, which is great for time but
not so great for space. Why? Step (2) produces sig context heaps that are consumed by step
(3). Reading large sig context heaps from shared memory causes a lot of allocation, which is
furthermore cached for reuse in the same bucket of work. When all cores do this, OOMs are likely.

(This used to be less of a problem in classic because of the reduced parallelism: while the same sig
context heaps were produced and consumed (read, cached), the biggest ones would hog a few cores
while other cores lay waiting.)

This diff reduces peak memory usage by reducing the bucket size. Thus every core gets less work at a
time, so it can read and cache less at a time. The smaller bucket size causes a small time
regression, but the peak memory reduction is comparatively larger and de-risks OOMs by a lot on
large codebases.

Reviewed By: nmote

Differential Revision: D13254317

fbshipit-source-id: f19089308fa3fd77897bb066f4d38fd7c629cf5a