disallow shapes appearing as upper bounds of non-object types
Summary:
This changes the fall-through behavior of ShapeTs in upper bound positions to produce an assignment from EmptyT rather than AnyT. This means that `null` and `void` are no longer subtypes of Shapes.

This prevents behavior like this: https://flow.org/try/#0C4TwDgpgBAYg9nKBeKBvKBDKAuKA7AVwFsAjCAJygF8BuAKDoBsJgoAPHKAEgGUALDJAA88OAD5k+Ao0b02AOiwoAzDSA

This also updates the behavior of InstanceTs to match the behavior of objects when appearing as the lower bound of shapes, in that they should now ignore polarity. This is necessary since we no longer flow to `any` in this case.

Reviewed By: panagosg7

Differential Revision: D13201962

fbshipit-source-id: ac567c197df389a2d639a6a113982583be272549