Remove `offset` from ESTree translator
Summary:
In preparation to remove `offset` from locations for the memory savings, this diff migrates the ESTree translator to use `Offset_utils` instead.

This causes a change in test output due to different handling of multi-byte characters:

Currently, the `offset` field is based on number of characters, no matter their size in any particular character encoding.

Esprima reports offsets based on the number of 16-bit units when the string is encoded in UTF-16. This means that many common characters which have a multi-byte encoding in UTF-8 but have a 16-bit encoding in UTF-16 will not be treated any differently than ASCII characters. This is how indexing into a JS string behaves.

`Offset_utils` reports offsets in terms of number of bytes. It would also be fairly straightforward to add a mode to simulate what Esprima does, if that becomes necessary.

Reviewed By: gabelevi

Differential Revision: D13453396

fbshipit-source-id: 9459509a29b1e81b34fcc690644950356b5f60f3