Refactor Members module and associated functionality out of flow_js so its functions cannot be used during typechecking
Summary: These functions were only being used by the `get-def` and `autocomplete` language services, so this moves them out of `flow_js.ml` so that the code in that file cannot use them. This doesn't add any new functions to the `flow_js` interface, which seemed to be the concern around doing this originally.

Reviewed By: panagosg7

Differential Revision: D13522877

fbshipit-source-id: 4049eb0b4a77e669f01cf6a6c481f60520f2714b