minor refactoring and comments on non-lazy initialization
Summary: `files_to_infer` is only called in full check mode, with or without `focus_targets`. In follow up diffs I'll separate work done in this path with work done in other paths, so documenting what's expected along this path for now. Along with the documentation, I tweaked the input signature of `files_to_infer` to reflect the fact that it takes optional focus targets instead of (what sounds like a well-formed set of) "focused" files.

Reviewed By: nmote

Differential Revision: D13490194

fbshipit-source-id: d7a6ba9969378869093a78111d7f372a5a9c23e4