no types-first parse results in classic mode
Summary:
In classic mode, we were returning types-first parse results which contain more data than classic parse results and yet the additional data was never used. This caused a space regression.

This diff should revert that space regression. The additional space will have to pay for itself when types-first is rolled out, or be reduced. In particular, there is a lot of redundancy in `sig_file_sig` vs. `file_sig`. It is also conceivable that neither `ast` or  `sig_ast` needs to be stored; instead, we can store the contents and parse / derive on demand.

Reviewed By: nmote

Differential Revision: D13535173

fbshipit-source-id: eb154eed7bf4625b85959f3c6f666191fc02626a