WorkloadStream
Summary:
A workload is basically a command. Previously, `ServerMonitorListenerState.ml` had an `Lwt_stream.t` of workloads which it used to hold the workloads that still needed to be run. However, with parallelizable workloads we need a data structure that is a little more complicated.

Long story short we have

* nonparallelizable workloads (which need to be run by themselves after a recheck)
* parallelizable workloads (which can be run in parallel with a recheck)
* requeued parallelizable workloads (which can't be run until after the recheck finishes)

WorkloadStream is a data structure which lets us push workloads, pop the oldest workloads, and wait for workloads to be available.

Reviewed By: jbrown215

Differential Revision: D13511802

fbshipit-source-id: c409b44e36498e9008a5d0abc9ada0b582055572