ServerEnv.env ref -> ServerEnv.env
Summary:
I think we originally added the `ServerEnv.env ref` back when we first built lazy mode. When a command would run `ensure_checked_dependencies`, it might do a recheck and then update the ref.

Nowadays, `ensure_checked_dependencies` doesn't actually need to update the `env` (it just enqueues the files to be rechecked and cancels itself). The only commands which actually need to update the `env` are `find-refs` (which does lots of forcing itself) and `refactor`, which uses `find-refs`.

So rather than using a ref everywhere, I'm removing the ref from almost everywhere. This makes it clear which commands are actually mutating the `env`.

Reviewed By: nmote

Differential Revision: D13511800

fbshipit-source-id: 0b9048ae34584f11839e26e4b81e1e3cc05cb5d6