Don't cancel rechecks for redundant forcings
Summary:
Consider a world where we can run commands like `autocomplete` in parallel with a recheck. What happens if the server is in lazy mode, needs to check some dependencies, and `autocomplete` commands keep getting sent?

1 - We start a recheck
2 - In parallel we run an `autocomplete`
3 - It notices we need to check the dependency `foo`, cancels itself (and the current recheck) and asks us to force `foo` to be a dependency
4 - We start a new recheck to force `foo` to be a dependency
2 - In parallel we run an `autocomplete`
3 - It notices we need to check the dependency `foo`, cancels itself (and the current recheck) and asks us to force `foo` to be a dependency
4 - We start a new recheck to force `foo` to be a dependency
2  In parallel we run an `autocomplete`
...etc

We need to prevent rechecks from being canceled and restarted when they're already doing the work that we require.

This diff filters `files_to_force` and removes files which are already in `env.checked_files` or are currently being rechecked.

Reviewed By: jbrown215

Differential Revision: D13511798

fbshipit-source-id: 2b29f33186e2ece0d11a38b6fe15b1aebcf4e036