[refactor] Make the slab in flow_js.ml a module
Summary:
Turns the Slab into a module. In a stacked diff, this module becomes a functor that is parameterized by React_kit.

This is part of a larger effort to reduce the complexity of flow_js.ml. I am going to be putting up a lot of diffs that refactor pieces of functionality into separate files to make things easier to digest. All of these files will have functors parameterized by Flow_common.S (created later in the stack), which is a signature that exposes several functions in the M__flow slab.

The first success I found was successfully getting rid of the clowny "passing functions between react_kit.ml and flow_js.ml" in order to avoid a dependency cycle. You can find this diff higher up in the stack.

Reviewed By: gabelevi

Differential Revision: D13532238

fbshipit-source-id: 773601b3ceaf7000f2c1550c7e3ce1dcd265db7e