[refactor] Functorize flow_js.ml and react_kit.ml
Summary: This turns the flow slab module into a functor that depends on the REACT signature. Similarly, this turns the React_kit.Kit module into a functor that depends on Flow_common.S. This sets us up to get rid of the way that we currently handle React_kit.run, which passes several functions from the slab into react_kit.ml.

Reviewed By: gabelevi

Differential Revision: D13532237

fbshipit-source-id: 334e7cc8a291e96fa6b60d513ccf1065b00063e2