use non-dynamic types to create iterators for for-of loops
Summary: Previously we created `any`s to fill in the `Return` and `Next` parameters of the `$Iterator` type constructor, which unnecessarily would have propagated taint to these values in any structure being iterated over. These parts of the iterator/generator are not ever used during `for ... of` iteration, so we can just have them be `mixed` and `empty`, respectively (since `Return` is covariant and `Next` is contravariant).

Reviewed By: panagosg7

Differential Revision: D13495572

fbshipit-source-id: 2982b5ebcc20632b8857e4260bb22b87f79cb261