Un-trust RTypeAlias after substitution
Summary:
Before in:
```
declare function map<TNext>(fn: O<TNext> | TNext): TNext;
declare var p: Poly<number>;
const p1 = map(p);
```
Flow type-at-pos would infer the type of `p1` as
```
Poly
```
instead of
```
Poly<number>
```

This had to do with the way the normalizer reconstructs type aliases.
Whenever a type has an RTypeAlias reason, then this is used to recover the name
of the type alias, instead of its type expansion. This mechanism, however, broke down
when polymorphic instances were instantiated through `Subst.subst`. There, the exact
same reason was preserved post-substitution. This would cause an ill-formed type to
be reconstructed since after the substitution, the type is not guarder by a TypeAppT
anymore.

This diff fixes this by flipping the "trust" bit of the RTypeAlias reason to false after a
substitution is performed. In the above example the result is now:
```
{ f: number }
```

Reviewed By: avikchaudhuri

Differential Revision: D13525276

fbshipit-source-id: 5d3f3ce2223d7b2ee1d33cc4ee61a9caec761a45