remove Loc.contains call in Flow_js
Summary:
The `Loc.contains` call was blocking the abstract location conversion project, so it needed to be replaced with something that replicated its behavior. Rather than calculating alternate `use_op`s based on the location of one containing the other, we now use a heuristic based on the `use_op`s themselves to approximate the relative locality of the `use_op` to the error.

The approximation isn't 100% perfect; it moves the error reporting location slightly for some polymorphic functions. However, the change is solely in the error location, it is not semantic. Additionally, given that the original implementation was a self-described hack, I'm not entirely convinced that perfectly mimicking the original behavior is a goal at all.

Reviewed By: panagosg7

Differential Revision: D13542665

fbshipit-source-id: 602fc5ea1816325cb208d1a23891079c353211ef