[PR] Add `[untyped]` documentation
Summary:
[This commit](https://github.com/facebook/flow/commit/1fdf33de5e46d9cf15fd5cdae9e7d53e27594d44) introduced the `[untyped]` config section.

As in https://github.com/facebook/flow/issues/6749 this is present and is suggested to be used in several issues, but is not documented.

This is a first pass at adding documentation for it. I only started looking at Flow yesterday, so hopefully it is OK! �
Pull Request resolved: https://github.com/facebook/flow/pull/7129

Reviewed By: jbrown215

Differential Revision: D12945646

Pulled By: nmote

fbshipit-source-id: 12a79306d4d6cffc9158537ece1734e2aaacbbd6