Use newer version of recast
Summary:
We used to rely on a broken version of recast to do our react codemods. Since it was fixed, we can update that codemod to no longer use the broken version. This is also necessary because our new codemods rely on a newer version of recast.

Since recast still has a bug in it when it traverses opaque types, we're using a fork that addresses this issue.
Also, recast doesn't know about the inexact syntax yet, so that fork is double-ly necessary.

Note also that the pretty printer started to surround class expressions in parens, so tests had to be re-recorded

Reviewed By: gabelevi

Differential Revision: D10853869

fbshipit-source-id: be50f11cd3c4a87f795dfc8c79f7daa8bcc62eca