Fix opaque_id and class_id comparison in Flow_js
Summary:
An opaque type (resp. class instance type) is identified by the line and column span of
it's name at the point of definition. Across rechecks we consider the name of an opaque
type to have remained invariant iff the line and column span remains the same (regardless
of any discrepancies to the offset) -- see `add_loc` in sigHash.ml. This diff copies this logic
to the part of the checker that compares two opaque types.

`loc1 = loc2` is not equivalent to `ALoc.equal loc1 loc2` due to `offset`.

This caused bogus recheck errors to appear whenever the contents of a the file of a
remote opaque type changed, but not the line and column of the opaque type definition.

Reviewed By: avikchaudhuri

Differential Revision: D13614013

fbshipit-source-id: 6a775df51ea4f85799d34fd6223ad83e51699678