Use offset utils for JSON output of locations
Summary:
I was initially going to remove offset from JSON output altogether, but Gabe Levi convinced me to delay the breaking change. The plan now is to leave offset in the output using offset utils (this diff), then introduce a flag (default on) to include offset output, then after one release change it to default off, then remove it altogether.

In a few places, most notably error output, this diff uses the super clowny approach of repeatedly reading the entire contents of a file in order to compute offsets. At lest for error output, I was initially planning to write some sort of cache to store file contents and offset tables, but I ran some performance experiments and it turned out to just not really matter. The time taken to render ~600 errors on my devserver varied between 1.0 and 1.1 seconds both before and after this change. Maybe we'll need to do something better at some point but for now, the simple approach seems adequate.

One issue is that if files change on disk between the time when locations are constructed and when we compute offsets, we could end up reporting bogus offsets. This could also cause us to omit offsets altogether when there is no entry in the offset table for a given position (we elect to omit the offset instead of crashing because most consumers probably won't care about the offset).

Reviewed By: gabelevi

Differential Revision: D13600021

fbshipit-source-id: f14543eb3040837d6a7ef140de6c9e2364f89c47