[Typed AST] Faithful error nodes
Summary:
This diff makes error nodes in the typed AST structurally equivalent to their input counterparts.

One of the main sources of unfaithfulness of the typed AST compared to the input ASTs
was the handling of errors nodes. Before, a typical pattern for error nodes was to replace
the subtree rooted at the source of the error with a generic error node. So assuming the
input AST:
```
       A
     /   \
   B       C
  /  \
D      E
```
If Flow found an error on `B` then it would produce the following typed AST:
```
       A
     /   \
   Error   C
```
where `Error` is a generic error node.

The main downside was that this considerably distorted the form of the AST making it harder to:
* test the construction of the typed AST
* reliably port services like coverage (in the example above the total number of nodes
would change)

This changes the above behavior by constructing a typed AST that respects the input AST
structure. In the example above we would produce:
```
       A
     /   \
   B'      C
  /  \
D'     E'
```
where the primed nodes are structural replicas of their input counterparts, but where the
type annotations are indicative of the attendant error.

The type annotation on these nodes will be AnyT with one of the following `unsoundness_kind`s:
* Error: typical Flow error.
* Unimplemented: nodes marked as TODO in the typechecker.
* Unchecked: nodes deliberately (?) left unchecked (e.g. ` <element name={} /> `).

Reviewed By: samwgoldman

Differential Revision: D10123488

fbshipit-source-id: 7eb103668bca42382b7d581fa8d391d2a1712916